arrumando os bindings
